+++ frontend/controllers/FavoritoController.php
@@ -38,13 +38,18 @@ class FavoritoController extends Controller
      */
     public function actionIndex()
     {
-        $db_favoritos = Favorito::find()
-            ->where(['id_user' => Yii::$app->user->id])
-            ->all();
-    
-        return $this->render('index', [
-            'db_favoritos' => $db_favoritos,
-        ]);
+        if (Yii::$app->user->can('visualizarFavoritos')) {
+            $db_favoritos = Favorito::find()
+                ->where(['id_user' => Yii::$app->user->id])
+                ->all();
+
+            return $this->render('index', [
+                'db_favoritos' => $db_favoritos,
+            ]);
+        }else{
+            Yii::$app->session->setFlash('danger', ' Não têm permissões para visualizar os favoritos');
+            return $this->redirect(['site/index']);
+        }
     }
 
     /**
@@ -54,14 +59,20 @@ class FavoritoController extends Controller
      */
     public function actionCreate($codigo_produto)
     {
-        $model_favorito = new Favorito();
-    
-        $model_favorito->codigo_produto = $codigo_produto;
-        $model_favorito->id_user = Yii::$app->user->getId();
-        $model_favorito->save();
-    
-        Yii::$app->session->setFlash('success', $model_favorito->produto->modelo->nome . ' ' . $model_favorito->produto->nome . ' foi adicionado à sua lista de favoritos.');
-        return $this->redirect(Yii::$app->request->referrer);
+        if (Yii::$app->user->can('adicionarProdutosFavoritos')) {
+
+            $model_favorito = new Favorito();
+
+            $model_favorito->codigo_produto = $codigo_produto;
+            $model_favorito->id_user = Yii::$app->user->getId();
+            $model_favorito->save();
+
+            Yii::$app->session->setFlash('success', $model_favorito->produto->modelo->nome . ' ' . $model_favorito->produto->nome . ' foi adicionado à sua lista de favoritos.');
+            return $this->redirect(Yii::$app->request->referrer);
+        }else{
+            Yii::$app->session->setFlash('danger', ' Não têm permissões para adicionar aos favoritos');
+            return $this->redirect(['site/index']);
+        }
     }
 
     /**
@@ -73,11 +84,16 @@ class FavoritoController extends Controller
      */
     public function actionDelete($id)
     {
-        $model_favorito = $this->findModel($id);
-        $model_favorito->delete();
-    
-        Yii::$app->session->setFlash('danger', $model_favorito->produto->modelo->nome . ' ' . $model_favorito->produto->nome . ' foi removido da sua lista de favoritos.');
-        return $this->redirect(Yii::$app->request->referrer);
+        if (Yii::$app->user->can('eliminarFavoritos')) {
+            $model_favorito = $this->findModel($id);
+            $model_favorito->delete();
+
+            Yii::$app->session->setFlash('danger', $model_favorito->produto->modelo->nome . ' ' . $model_favorito->produto->nome . ' foi removido da sua lista de favoritos.');
+            return $this->redirect(Yii::$app->request->referrer);
+        }else {
+            Yii::$app->session->setFlash('danger', ' Não têm permissões para eliminar favoritos');
+            return $this->redirect(['site/index']);
+        }
     }
     
     /**
